+++ docs/sidebars.js
@@ -35,7 +35,14 @@ module.exports = {
       type: 'category',
       label: 'API Overview',
       collapsed: false,
-      items: ['api/observable', 'api/action', 'api/reaction', 'api/observers', 'api/context', 'api/spy'],
+      items: [
+        'api/observable',
+        'api/action',
+        'api/reaction',
+        'api/observers',
+        'api/context',
+        'api/spy',
+      ],
     },
     {
       Development: ['development/blueprint', 'development/release'],