+++ mobx/lib/mobx.dart
@@ -64,5 +64,4 @@ export 'package:mobx/src/core.dart'
         DerivationState;
 export 'package:mobx/src/core/atom_extensions.dart';
 
-/// The current version as per `pubspec.yaml`
-const version = '2.0.7+5';
+export 'version.dart';